+++ ServicesManagementCentral.Web/Models/Cotizador/DataResponseModel.cs
@@ -14,11 +14,11 @@ namespace ServicesManagement.Web.Models.Cotizador
         public string deliveryEstimate { get; set; }
         public DeliveryDateModel deliveryDate { get; set; }
         public int quantity { get; set; }
-        public int basedPrice { get; set; }
+        public decimal basedPrice { get; set; }
         public int extendedFare { get; set; }
         public int insurance { get; set; }
         public int additionalServices { get; set; }
-        public int totalPrice { get; set; }
+        public decimal totalPrice { get; set; }
         public string currency { get; set; }
         public bool customKey { get; set; }
         public int importFree { get; set; }
